+++ TG.js
@@ -1,9 +1,9 @@
 // combine the user-supplied options with the referencing base options
-function combine(object, options) {
+function combine(object, options, whitelistedKeys) {
   Object.keys(object).forEach((key) => {
     if (object[key] instanceof Object) {
-      combine(object[key], options)
-    } else if (options[key] !== undefined) {
+      combine(object[key], options, whitelistedKeys)
+    } else if (options[key] !== undefined && whitelistedKeys.indexOf(key) >= 0) {
       object[key] = options[key]
     }
   })
@@ -15,34 +15,40 @@ class TG {
   }
 
   /*
-   *  Sends text message to a existing chat.
+   *  Send text message to an existing chat.
    *  Method "parseTextEntities" requires TDLib 1.1.0 (git.io/tdlibchanges).
    */
-  async sendTextMessage(chatId, text, options = {}) {
-    let formattedText
-    switch (options.parse_mode) {
-      case 'html':
-        formattedText = await this.client._execute({
-          '@type': 'parseTextEntities',
-          'parse_mode': { '@type': 'textParseModeHTML' },
-          text,
-        })
-        break
-      case 'markdown':
-        formattedText = await this.client._execute({
-          '@type': 'parseTextEntities',
-          'parse_mode': { '@type': 'textParseModeMarkdown' },
-          text,
-        })
-        break
-      default: {
-        formattedText = {
-          '@type': 'formattedText',
-          text,
-        }
-        break
-      }
+  async sendTextMessage(args = {}) {
+    const { $text, ...options } = args
+    if (!$text) {
+      throw new Error('No text defined for method "sendTextMessage".')
     }
+    const payload = {
+      '@type': 'sendMessage',
+      'chat_id': 0,
+      'reply_to_message_id': 0,
+      'disable_notification': false,
+      'from_background': true,
+      'reply_markup': null,
+      'input_message_content': {
+        '@type': 'inputMessageText',
+        'text': await $text._format(this.client),
+        'disable_web_page_preview': true,
+        'clear_draft': true,
+      },
+    }
+    combine(payload, options, [
+      'chat_id', 'reply_to_message_id', 'disable_notification', 'from_background',
+      'disable_web_page_preview', 'clear_draft',
+    ])
+    return this.client.fetch(payload)
+  }
+
+  /*
+   *  Send photo message to an existing chat.
+   */
+  async sendPhotoMessage(args = {}) {
+    const { $caption, ...options } = args
     const payload = {
       '@type': 'sendMessage',
       'chat_id': chatId,
@@ -51,13 +57,23 @@ class TG {
       'from_background': true,
       'reply_markup': null,
       'input_message_content': {
-        '@type': 'inputMessageText',
-        'text': formattedText,
-        'disable_web_page_preview': true,
-        'clear_draft': true,
+        '@type': 'inputMessagePhoto',
+        'photo': {
+          '@type': 'inputFileLocal',
+          'path': path,
+        },
+        'thumbnail': null,
+        'added_sticker_file_ids': [],
+        'width': 0,
+        'height': 0,
+        'caption': $caption ? await $caption._format(this.client) : null,
+        'ttl': 0,
       },
     }
-    combine(payload, options)
+    combine(payload, options, [
+      'chat_id', 'reply_to_message_id', 'disable_notification', 'from_background',
+      'path', 'thumbnail', 'added_sticker_file_ids', 'width', 'height', 'ttl',
+    ])
     return this.client.fetch(payload)
   }