+++ structs/index.js
@@ -1,0 +1,3 @@
+const TextStruct = require('./TextStruct')
+
+module.exports = { TextStruct }